# src/docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"_36_Enums.TipoUser": {
				"type": "string",
				"enum": [
					"DOCENTE",
					"TI",
					"ADM"
				]
			},
			"TipoUser": {
				"$ref": "#/components/schemas/_36_Enums.TipoUser"
			},
			"CreateUser": {
				"properties": {
					"email": {
						"type": "string"
					},
					"tipoUser": {
						"$ref": "#/components/schemas/TipoUser"
					}
				},
				"required": [
					"email",
					"tipoUser"
				],
				"type": "object",
				"additionalProperties": true
			},
			"_36_Enums.StatusConta": {
				"type": "string",
				"enum": [
					"DESATIVADA",
					"ATIVA",
					"CONVIDADA"
				]
			},
			"StatusConta": {
				"$ref": "#/components/schemas/_36_Enums.StatusConta"
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "backend_pi_4_semestre",
		"version": "1.0.0",
		"description": "1. Inicialização e Estrutura de Pastas\r Abra o terminal na pasta do projeto e execute:",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Jhon"
		}
	},
	"paths": {
...
		"/Coordenacao/Criar_Usuario": {
			"post": {
				"operationId": "HandleCreateUser",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"400": {
						"description": "Validation Failed"
					}
				},
				"tags": [
					"coordenador"
				],
				"security": [
					{
						"jwt": [
							"ADM"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateUser"
							}
						}
					}
				}
			}
		},
		"/Coordenacao/listar": {
			"get": {
				"operationId": "HandleFindUsers",
				"responses": {
					"200": {
						"description": "Search Completed",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"msg": {
											"type": "string"
										}
									},
									"required": [
										"msg"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"coordenador"
				],
				"security": [
					{
						"jwt": [
							"ADM"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "tipoUser",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/TipoUser"
						}
					},
					{
						"in": "query",
						"name": "statusConta",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/StatusConta"
						}
					}
				]
			}
		},
		"/Coordenacao/desativa-user": {
			"patch": {
				"operationId": "HandleDesativarUser",
				"responses": {
					"200": {
						"description": "Sucesso ao alterar status do usuário",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"msg": {
											"type": "string"
										}
									},
									"required": [
										"msg"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"coordenador"
				],
				"security": [
					{
						"jwt": [
							"ADM"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "idUser",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "statusUser",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/StatusConta"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}